rxk5-1557-20090510
LICENSE IPL10
FIXES n/a

Update rxk5 branch to 1.5.59.

+++ src/afs/LINUX/osi_file.c
@@ -64,7 +64,7 @@ osi_UFSOpen(afs_int32 ainode)
     AFS_GUNLOCK();
     if (!afile) {
 	osi_Panic("osi_UFSOpen: Failed to allocate %d bytes for osi_file.\n",
-		  sizeof(struct osi_file));
+		  (int)sizeof(struct osi_file));
     }
     memset(afile, 0, sizeof(struct osi_file));
 #if defined(HAVE_IGET)
